refactor(action-bars): update
