+++ src/containers/warp/components/buttonGroup/stylesBtn.scss
@@ -97,7 +97,7 @@ $color-border-lamp-white: #e2e2e2;
   outline: none;
   border: none;
   background: transparent;
-  cursor: pointer;
+  text-transform: lowercase;
 
   &:hover {
     opacity: 0.8;
