+++ src/containers/warp/type.ts
@@ -15,3 +15,9 @@ export interface PoolsWithAssetsCapType extends PoolsWithAssetsType {
 export type OptionNeverArray<T> = T | never[];
 
 export type TypeTab = 'create-pool' | 'add-liquidity' | 'sub-liquidity';
+
+export const enum TypeTabEnum {
+  ADD_LIQUIDITY = 'add-liquidity',
+  CREATE_POOL = 'create-pool',
+  SUB_LIQUIDITY = 'sub-liquidity',
+}
