+++ src/layouts/Main.tsx
@@ -1,6 +1,7 @@
 import { useState } from 'react';
 import { useSelector } from 'react-redux';
-import { Telegram, GitHub } from 'src/components/actionBar';
+import { Telegram } from 'src/components/actionBar/Telegram';
+import { GitHub } from 'src/components/actionBar/GitHub';
 import AppMenu from 'src/containers/application/AppMenu';
 import AppSideBar from 'src/containers/application/AppSideBar';
 import Header from 'src/containers/application/Header/Header';
